+++ src/deprecated/InterruptHandlerComponentPkg_c.vhd
@@ -63,11 +63,11 @@ package InterruptHandlerComponentPkg is
     IntReq       : in   std_logic_vector(gIntReq'range) ;
 
     -- From TestCtrl
-    TransRec     : view AddressBusVerificationComponentView of AddressBusRecType ;
-    InterruptRec : view AddressBusVerificationComponentView of AddressBusRecType ;
+    TransRec     : inout AddressBusRecType ;
+    InterruptRec : inout AddressBusRecType ;
     
     -- To Verification Component
-    VCRec        : view AddressBusTestCtrlView of AddressBusRecType
+    VCRec        : inout AddressBusRecType
   ) ;
   end component InterruptHandler ;